+++ baus/datasources.py
@@ -5,7 +5,7 @@ from urbansim_defaults import datasources
 from urbansim_defaults import utils
 from urbansim.utils import misc
 import orca
-import allocation
+import preprocessing
 from utils import geom_id_to_parcel_id, parcel_id_to_geom_id
 from utils import nearest_neighbor
 
@@ -113,54 +113,10 @@ def building_sqft_per_job(settings):
 
 
 # key locations in the Bay Area for use as attractions in the models
-@orca.table('landmarks', cache=True)
+@orca.table(cache=True)
 def landmarks():
     return pd.read_csv(os.path.join(misc.data_dir(), 'landmarks.csv'),
                        index_col="name")
-
-
-def move_jobs_from_portola_to_san_mateo_county(parcels, buildings, jobs_df):
-    # need to move jobs from portola valley to san mateo county
-    NUM_IN_PORTOLA = 1500
-
-    jobs_df["parcel_id"] = misc.reindex(buildings.parcel_id,
-                                        jobs_df.building_id)
-    jobs_df["juris"] = misc.reindex(parcels.juris, jobs_df.parcel_id)
-
-    portola = jobs_df[jobs_df.juris == "Portola Valley"]
-    move = portola.sample(len(portola) - NUM_IN_PORTOLA)
-
-    san_mateo = jobs_df[jobs_df.juris == "San Mateo County"]
-    move_to = san_mateo.sample(len(move))
-
-    jobs_df.loc[move.index, "building_id"] = move_to.building_id.values
-
-    return jobs_df
-
-
-@orca.table(cache=True)
-def jobs(store, parcels, recompute_cached_tables):
-
-    if 'jobs_cached' in store and not recompute_cached_tables:
-        return store['jobs_cached']
-
-    if 'jobs_urbansim_allocated' not in store:
-        # if jobs allocation hasn't been done, then do it
-        # (this should only happen once)
-        orca.run(["allocate_jobs"])
-
-    jobs_df = store['jobs_urbansim_allocated']
-
-    # grab it from store in order to avoid circular reference - if you
-    # use the orca buildings table it depends on jobs so in jobs we can't
-    # depend on buildings
-    buildings = store.buildings
-    jobs_df = move_jobs_from_portola_to_san_mateo_county(
-        parcels, buildings, jobs_df)
-
-    store['jobs_cached'] = jobs_df
-
-    return jobs_df
 
 
 @orca.table(cache=True)
@@ -177,7 +133,7 @@ def base_year_summary_taz():
 
 
 # non-residential rent data
-@orca.table('costar', cache=True)
+@orca.table(cache=True)
 def costar(store, parcels):
     df = pd.read_csv(os.path.join(misc.data_dir(), '2015_08_29_costar.csv'))
     df["PropertyType"] = df.PropertyType.replace("General Retail", "Retail")
@@ -197,14 +153,8 @@ def costar(store, parcels):
 
 @orca.table(cache=True)
 def zoning_lookup():
-    df = pd.read_csv(os.path.join(misc.data_dir(), "zoning_lookup.csv"))
-    # this part is a bit strange - we do string matching on the names of zoning
-    # in order to link parcels and zoning and some of the strings have small
-    # differences, so we copy the row and have different strings for the same
-    # lookup row.  for now we drop duplicates of the id field in order to run
-    # in urbansim (all the attributes of rows that share an id are the same -
-    # only the name is different)
-    return df.drop_duplicates(subset='id').set_index('id')
+    df = pd.read_csv(os.path.join(misc.data_dir(), "zoning_lookup.csv"),
+                     index_col='id')
 
 
 # zoning for use in the "baseline" scenario
@@ -218,8 +168,8 @@ def zoning_baseline(parcels, zoning_lookup, settings):
                   left_on="zoning_id", right_index=True)
     df = geom_id_to_parcel_id(df, parcels)
 
+    # change column names
     d = {k: "type%d" % v for k, v in settings["building_type_map2"].items()}
-
     df.columns = [d.get(x, x) for x in df.columns]
 
     return df
@@ -393,127 +343,18 @@ def development_projects(parcels, settings, scenario):
 
 
 @orca.table(cache=True)
-def households(store, settings, parcels_geography, recompute_cached_tables):
-
-    if 'households_cached' in store and not recompute_cached_tables:
-        return store['households_cached']
-
-    # start with households from urbansim_defaults
-    df = datasources.households(store, settings)
-
-    # need to keep track of base year income quartiles for use in the
-    # transition model - even caching doesn't work because when you add
-    # rows via the transitioning, you automatically clear the cache!
-    # this is pretty nasty and unfortunate
-    df["base_income_quartile"] = pd.Series(pd.qcut(df.income, 4, labels=False),
-                                           index=df.index).add(1)
-    df["base_income_octile"] = pd.Series(pd.qcut(df.income, 8, labels=False),
-                                         index=df.index).add(1)
-
-    # there are some overrides where we move households around in order
-    # to match the city totals - in the future we will resynthesize and this
-    # can go away - this csv is generated by scripts/match_city_totals.py
-    overrides = pd.read_csv("data/household_building_id_overrides.csv",
-                            index_col="household_id").building_id
-    df.loc[overrides.index, "building_id"] = overrides.values
-
-    # turns out we need 4 more households
-    new_households = df.loc[
-        [1132542, 1306618, 950630, 886585]].reset_index()
-    # keep unique index
-    new_households.index += pd.Series(df.index).max() + 1
-    df = df.append(new_households)
-
-    store['households_cached'] = df
-
-    return df
+def jobs(store):
+    return store['jobs_preproc']
 
 
 @orca.table(cache=True)
-def buildings(store, parcels, households, jobs, building_sqft_per_job,
-              settings, manual_edits, recompute_cached_tables):
+def households(store):
+    return store['households_preproc']
 
-    if 'buildings_cached' in store and not recompute_cached_tables:
-        return store['buildings_cached']
 
-    # start with buildings from urbansim_defaults
-    df = datasources.buildings(store, households, jobs,
-                               building_sqft_per_job, settings)
-
-    df = df.drop(['development_type_id', 'improvement_value',
-                  'sqft_per_unit', 'nonres_rent_per_sqft',
-                  'res_price_per_sqft',
-                  'redfin_home_type', 'costar_property_type',
-                  'costar_rent'], axis=1)
-
-    edits = manual_edits.local
-    edits = edits[edits.table == 'buildings']
-    for index, row, col, val in \
-            edits[["id", "attribute", "new_value"]].itertuples():
-        df.set_value(row, col, val)
-
-    # set the vacancy rate in each building to 5% for testing purposes
-    df["residential_units"] = df.residential_units.fillna(0)
-
-    # for some reason nonres can be more than total sqft
-    df["building_sqft"] = pd.DataFrame({
-        "one": df.building_sqft,
-        "two": df.residential_sqft + df.non_residential_sqft}).max(axis=1)
-
-    # keeps parking lots from getting redeveloped
-    df["building_sqft"][df.building_type_id.isin([15, 16])] = 0
-    df["non_residential_sqft"][df.building_type_id.isin([15, 16])] = 0
-
-    # don't know what a 0 building type id, set to office
-    df["building_type_id"] = df.building_type_id.replace(0, 4)
-
-    df["redfin_sale_year"] = df.redfin_sale_year.fillna(2012)
-
-    # hope we get more data on this soon
-    df["deed_restricted_units"] = 0
-
-    if "skip_deed_restricted_units" in settings:
-        return df
-
-    zone_ids = misc.reindex(parcels.zone_id, df.parcel_id).\
-        reindex(df.index).fillna(-1)
-
-    # sample deed restricted units to match current deed restricted unit
-    # zone totals
-    for taz, row in pd.read_csv('data/deed_restricted_zone_totals.csv',
-                                index_col='taz_key').iterrows():
-
-        cnt = row["units"]
-
-        if cnt <= 0:
-            continue
-
-        potential_add_locations = df.residential_units[
-            (zone_ids == taz) &
-            (df.residential_units > 0)]
-
-        assert len(potential_add_locations) > 0
-
-        weights = potential_add_locations / potential_add_locations.sum()
-
-        buildings_ids = potential_add_locations.sample(
-            cnt, replace=True, weights=weights)
-
-        units = pd.Series(buildings_ids.index.values).value_counts()
-        df.loc[units.index, "deed_restricted_units"] += units.values
-
-    print "Total deed restricted units after random selection: %d" % \
-        df.deed_restricted_units.sum()
-
-    df["deed_restricted_units"] = \
-        df[["deed_restricted_units", "residential_units"]].min(axis=1)
-
-    print "Total deed restricted units after truncating to res units: %d" % \
-        df.deed_restricted_units.sum()
-
-    store['buildings_cached'] = df
-
-    return df
+@orca.table(cache=True)
+def buildings(store):
+    return store['buildings_preproc']
 
 
 @orca.table(cache=True)
@@ -535,7 +376,7 @@ def household_controls(household_controls_unstacked):
     return df
 
 
-@orca.table('employment_controls_unstacked', cache=True)
+@orca.table(cache=True)
 def employment_controls_unstacked():
     df = pd.read_csv(os.path.join(misc.data_dir(), "employment_controls.csv"))
     return df.set_index('year')
