+++ baus/models.py
@@ -1013,92 +1013,3 @@ def price_vars(net):
     nodes = orca.get_table('nodes')
     nodes = nodes.to_frame().join(nodes2)
     orca.add_table("nodes", nodes)
-
-
-@orca.step()
-def correct_baseyear_data(buildings, parcels, jobs):
-    # sonoma county has too much vacancy in the buildings so we're
-    # going to lower it a bit to match job totals - I'm doing it here
-    # as opposed to in datasources as it requires registered orca
-    # variables
-
-    '''
-    These are the original vacancies
-    Alameda          0.607865
-    Contra Costa     0.464277
-    Marin            0.326655
-    Napa             0.427900
-    San Francisco    0.714938
-    San Mateo        0.285090
-    Santa Clara      0.368031
-    Solano           0.383663
-    Sonoma           0.434263
-    '''
-
-    # get buildings by county
-    buildings_county = misc.reindex(parcels.county, buildings.parcel_id)
-    buildings_juris = misc.reindex(parcels.juris, buildings.parcel_id)
-
-    # making sure we have no more than 10% vacancy
-    # this is the maximum vacancy you can have any a building so it NOT the
-    # same thing as setting the vacancy for the entire county
-    SURPLUS_VACANCY_COUNTY = buildings_county.map({
-       "Alameda": .42,
-       "Contra Costa": .57,
-       "Marin": .28,
-       "Napa": .7,
-       "San Francisco": .08,
-       "San Mateo": .4,
-       "Santa Clara": .32,
-       "Solano": .53,
-       "Sonoma": .4
-    }).fillna(.2)
-
-    SURPLUS_VACANCY_JURIS = buildings_juris.map({
-       "Berkeley": .65,
-       "Atherton": 0.05,
-       "Belvedere": 0,
-       "Corte Madera": 0,
-       "Cupertino": .1,
-       "Healdsburg": 0,
-       "Larkspur": 0,
-       "Los Altos Hills": 0,
-       "Los Gatos": 0,
-       "Monte Sereno": 0,
-       "Piedmont": 0,
-       "Portola Valley": 0,
-       "Ross": 0,
-       "San Anselmo": 0,
-       "Saratoga": 0,
-       "Woodside": 0,
-       "Alameda": .2
-    })
-
-    SURPLUS_VACANCY = pd.DataFrame([
-       SURPLUS_VACANCY_COUNTY, SURPLUS_VACANCY_JURIS]).min()
-
-    # count of jobs by building
-    job_counts_by_building = jobs.building_id.value_counts().\
-        reindex(buildings.index).fillna(0)
-    # with a 10% vacancy
-    job_counts_by_building_surplus = \
-        (job_counts_by_building * (SURPLUS_VACANCY+1)).astype('int')
-    # min of job spaces and 10% greater than number of jobs
-    correct_job_spaces = pd.DataFrame([
-        job_counts_by_building_surplus, buildings.job_spaces]).min()
-    # convert back to non res sqft because job spaces is computed
-    correct_non_res_sqft = correct_job_spaces * buildings.sqft_per_job
-
-    buildings.update_col("non_residential_sqft", correct_non_res_sqft)
-
-    jobs_county = misc.reindex(buildings_county, jobs.building_id)
-
-    print "Vacancy rate by county:\n", \
-        buildings.job_spaces.groupby(buildings_county).sum() / \
-        jobs_county.value_counts() - 1.0
-
-    jobs_juris = misc.reindex(buildings_juris, jobs.building_id)
-
-    s = buildings.job_spaces.groupby(buildings_juris).sum() / \
-        jobs_juris.value_counts() - 1.0
-    print "Vacancy rate by juris:\n", s.to_string()
